backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow==10.2.0
python-multipart==0.0.19
aiofiles==23.2.1
pydantic==2.5.2
numpy==1.26.4
ultralytics==8.1.7
psycopg2-binary==2.9.9
